% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moore.R
\docType{data}
\name{moore}
\alias{moore}
\title{Moore's Law data}
\format{A data frame with 48 rows and 3 variables:
\describe{
  \item{year}{year, from 1973 to 2011}
  \item{time}{time starting from 1973}
  \item{count}{number of transistors}
}}
\usage{
moore
}
\description{
A dataset containing the number of transistors and the corresponding #' years. The Moore's Law #'states that the number of transistors in a dense integrated circuit doubles approximately every two #'years.
moore.
}
\keyword{datasets}
